Функция открытия попапа перенесена в presenter
